Server/BackendSchluesselKasten/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
channels==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
colorzero==2.0
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.1
cssselect2==0.6.0
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-csp==3.7
django-select2==7.10.0
django-sortedm2m==3.1.1
djangorestframework==3.13.1
fonttools==4.33.3
future==0.18.2
gpiozero==1.6.2
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.9.1
msgpack==1.0.3
oscrypto==1.3.0
packaging==21.3
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.2.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.3
reportlab==3.6.10
requests==2.28.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
zope.interface==5.4.0
zopfli==0.2.1
